+++ help/index.html
@@ -1,8 +1,8 @@
 <!doctype html>
-<html lang="en-GB"><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1,viewport-fit=cover"><meta name="color-scheme" content="dark"><meta name="description" content="Guide for Toolkit v8.1.2 — The Godfather complete interface system, Offer payout and preserved v7 native product boundary."><title>MissionChief Map Command Toolkit — v8 Help Centre</title><style>
+<html lang="en-GB"><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1,viewport-fit=cover"><meta name="color-scheme" content="dark"><meta name="description" content="Guide for Toolkit v8.1.3 — The Godfather complete interface system, Offer payout and preserved v7 native product boundary."><title>MissionChief Map Command Toolkit — v8 Help Centre</title><style>
 :root{--bg:#061018;--panel:#0d2230;--line:#285066;--text:#edf8fc;--muted:#9db6c2;--cyan:#62d9ff;--gold:#ffd178}*{box-sizing:border-box}body{margin:0;background:linear-gradient(180deg,#061018,#0a1922);color:var(--text);font:15px/1.6 system-ui,-apple-system,"Segoe UI",sans-serif}header,main,footer{max-width:1180px;margin:auto;padding:22px}header{padding-top:48px}h1{font-size:clamp(38px,7vw,72px);line-height:1;margin:0}h2{margin-top:0;color:var(--cyan)}.tag{display:inline-block;padding:6px 10px;border:1px solid #6b572b;border-radius:999px;color:var(--gold)}.lead{max-width:900px;color:#c4d9e2;font-size:18px}.grid{display:grid;grid-template-columns:repeat(auto-fit,minmax(250px,1fr));gap:14px}.card,section{border:1px solid var(--line);border-radius:16px;background:rgba(13,34,48,.9);padding:20px}.card h3{margin:0 0 8px}.card p,li{color:#bfd0d8}.notice{border-color:#7f642b;background:#2c220d}.actions{display:flex;flex-wrap:wrap;gap:10px}.actions a{padding:10px 14px;border-radius:10px;background:#176f93;color:white;text-decoration:none;font-weight:700}code,kbd{background:#142d39;border:1px solid #34596a;border-radius:6px;padding:2px 6px}footer{color:var(--muted);text-align:center}@media(max-width:600px){header,main{padding:16px}header{padding-top:32px}}</style></head><body>
-<header><span class="tag">Guide for Toolkit v8.1.2 candidate</span><h1>The One We Knew Before</h1><p class="lead">Version 8 adds The Godfather as a complete interface system while preserving the v7 native MissionChief product boundary, retained command systems and deterministic teardown.</p><div class="actions"><a href="https://update.greasyfork.org/scripts/586018/MissionChief%20Map%20Command%20Toolkit.user.js">Install stable v7</a><a href="https://github.com/Conroy1988/missionchief-toolkit-assets/releases">Releases</a></div></header>
-<main><section class="notice"><h2>What changes in v8.1.2</h2><p>The Godfather becomes the eighth complete interface system, with original Toolkit-made artwork and a lazy theme-scoped Offer payout sting. Production remains on verified v7.1.6 until the guarded v8 release pipeline completes.</p></section>
+<header><span class="tag">Guide for Toolkit v8.1.3 candidate</span><h1>The One We Knew Before</h1><p class="lead">Version 8 adds The Godfather as a complete interface system while preserving the v7 native MissionChief product boundary, retained command systems and deterministic teardown.</p><div class="actions"><a href="https://update.greasyfork.org/scripts/586018/MissionChief%20Map%20Command%20Toolkit.user.js">Install stable v7</a><a href="https://github.com/Conroy1988/missionchief-toolkit-assets/releases">Releases</a></div></header>
+<main><section class="notice"><h2>What changes in v8.1.3</h2><p>The Godfather becomes the eighth complete interface system, with original Toolkit-made artwork and a lazy theme-scoped Offer payout sting. Production remains on verified v7.1.6 until the guarded v8 release pipeline completes.</p></section>
 <section><h2>Native Toolkit systems</h2><div class="grid">
 <div class="card"><h3>Mission Age map timers</h3><p>Compact age badges above personal missions. Press <kbd>6</kbd> to toggle them.</p></div>
 <div class="card"><h3>Mission Value</h3><p>Shows verified mission value in opened MissionChief windows without replacing native controls.</p></div>
@@ -31,4 +31,4 @@
         </ul>
       </section>
     </main>
-<footer>MissionChief Map Command Toolkit · v8.1.2 candidate · An Offer He Can’t Refuse</footer></body></html>
+<footer>MissionChief Map Command Toolkit · v8.1.3 candidate · An Offer He Can’t Refuse</footer></body></html>
